Add files via upload
Ping is a publicly available program that is used to check whether a host (or router/gateway) is reachable from the host. The operation of ping is very simple: ping uses the ICMP protocols mandatory ECHO REQUEST datagram to elicit an ICMP ECHO RESPONSE from a host or gateway. ECHO REQUEST datagrams (pings) have an IP and ICMP header, followed by a struct timeval and then an arbitrary number of pad bytes used to fill out the packet. Here a different version of ping program is presented.